chore(testing): move @playwright/test to devDependencies
The @playwright/test package is only required for running the test suite and is not needed for the production build of the application. For the web crawler, we will install `playwright` in the production environment.

Moving it from `dependencies` to `devDependencies` ensures that it will not be installed in the production environment, reducing the size of the final deployment package.